feat: use type submit on donate button
